runner: use the bash runfiles to resolve path
The gazelle_runner is a bash binary that invokes the gazelle golang binary
with some standard options. This bash binary needs to resolve
paths to gazelle, go and other plugins at runtime. It used a pretty simple
function to resolve paths. But that was too simple, and didn't work on windows.

This PR enables that function to be used on Windows as well.